5.7/Dockerfile: Add Azure Standard maintainer
Because we're not using generate-stackbrew-library.sh (which seems to
inject a MAINTAINER?).

I'm not sure what email address to use.  Jordan uses his personal
information for our Beehive and Redis containers, but it seems like we
should use something more Azure-wide if we're just pushing these to
azurestandard/ namespace.  Otherwise I expect we'd just stick with
Docker's library images.